digikam from trunk: Dimg fixes
- fix memory leak: added missing deref() in DImg::copy()
  The object that is copied would leak its memory
- fix crash in jpegloader

CCMAIL:digikam-devel@kde.org


svn path=/trunk/extragear/graphics/digikam/; revision=501049

+++ libs/dimg/dimgloader.cpp
@@ -40,7 +40,7 @@ int DImgLoader::granularity(DImgLoaderObserver *observer, int total, float progr
     // Progress slice is the part of 100% concerned with the current granularity
     // (E.g. in a loop only the values from 10% to 90% are used, then progressSlice is 0.8)
     // Current default is 1/20, that is progress info every 5%
-    int granularity = (total / (20 * progressSlice)) / observer->granularity();
+    int granularity = (int)(( total / (20 * progressSlice)) / observer->granularity());
     return granularity ? granularity : 1;
 }
